backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Yash Agrawal <yashagrawal01meh@gmail.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
psycopg = {extras = ["async"], version = "^3.2.3"}
bcrypt = "^4.2.1"
passlib = "^1.7.4"
argon2-cffi = "^23.1.0"
python-dotenv = "^1.0.1"
pydantic = "^2.10.2"
pydantic-settings = "^2.6.1"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
